修改dump问题，修改find，改为仅查找第一个，修改截图崩溃问题（仅捕获异常）

+++ src/autodriver/util/BlockDelayUtil.java
@@ -1,7 +1,7 @@
 package autodriver.util;
 
 public class BlockDelayUtil {
-	public static void runBlock(ExecuteBlock executeBlock, int timeout,Object... objects) {
+	public static void runBlock(ExecuteBlock executeBlock, float timeout, Object... objects) {
 		long starTime = System.currentTimeMillis();
 		while (executeBlock.excute(objects) == ExecuteBlock.StepResultWait && (System.currentTimeMillis() - starTime) < timeout * 1000) {
 			try {
@@ -12,7 +12,7 @@ public class BlockDelayUtil {
 		}
 	}
 
-	public static void waitforTime(int timeout) {
+	public static void waitforTime(float timeout) {
 		runBlock(new ExecuteBlock() {
 
 			@Override
